requirements.txt
Django==4.2.7
djangorestframework==3.14.0
PyMySQL==1.1.0
Pillow==10.1.0
django-cors-headers==4.3.1
python-decouple==3.8
redis==5.0.1
django-redis==5.4.0
celery==5.3.4
django-crispy-forms==2.3
crispy-bootstrap5==2025.6
requests==2.31.0
django-filter==23.5
reportlab==4.0.7
openpyxl==3.1.2
django-ckeditor==6.7.0
